pyproject.toml
[tool.poetry]
name = "estoque-api"
version = "0.1.0"
description = "Api para gerenciamento de estoque"
authors = ["Diniz <mamakerrykinh@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
